Initial test helper restructure
