+++ tests/_common.py
@@ -1,257 +1,57 @@
-"""Setup for tests for the beets-filetote plugin."""
+"""Backward-compatible shim for beets-filetote tests.
 
-import functools
+All implementations have moved to ``tests.pytest_beets_plugin``.
+This module re-exports them for existing unittest-based tests.
+Remove this file once all tests are migrated to native pytest style.
+"""
+
+# Re-export everything that existing tests import from here
+# Re-export the legacy TestCase for existing unittest-based tests
 import os
 import shutil
-import sys
 import tempfile
 import unittest
 
 from pathlib import Path
-from typing import TYPE_CHECKING
 
-from beets import config, logging, util
-from beets.util import FilesystemError
+from beets import config
 
-if TYPE_CHECKING:
-    from typing import TextIO
-
-# Propagate to root logger so nosetest can capture it
-log = logging.getLogger("beets")
-log.propagate = True
-log.setLevel(logging.DEBUG)
-
-PLATFORM = sys.platform
-
-# OS feature testing functions. These functions utilize `beets.util` functions
-# to test if symlinks, hardlinks, and reflinks are supported on the current platform
-# and filesystem, instead of just simple OS Platform checks (previously used).
-# These results are cached to avoid redundant checks across tests which helps
-# performance.
+from .pytest_beets_plugin._common import (  # noqa: F401
+    HAVE_HARDLINK,
+    HAVE_REFLINK,
+    HAVE_SYMLINK,
+    PLATFORM,
+    DummyIn,
+    DummyIO,
+    DummyOut,
+    InputError,
+    check_hardlink,
+    check_reflink,
+    check_symlink,
+)
+from .pytest_beets_plugin.assertions import AssertionsMixin
 
 
-@functools.lru_cache(maxsize=1)
-def check_symlink() -> bool:
-    """Tests if symlinks are usable using `beets.util.link`."""
-    with tempfile.TemporaryDirectory() as tmpdir_str:
-        tmpdir = Path(tmpdir_str)
-        src = tmpdir / "symlink_src"
-        dst = tmpdir / "symlink_dst"
-
-        try:
-            # Symlink source does not need to exist, it's just a path/pointer.
-            util.link(util.bytestring_path(src), util.bytestring_path(dst))
-
-            return dst.is_symlink()
-
-        except (FilesystemError, OSError) as e:
-            # Catch beets' custom error and any underlying OSError
-            log.debug(f"Symlink check failed: {e}")
-            return False
-
-
-@functools.lru_cache(maxsize=1)
-def check_hardlink() -> bool:
-    """Tests if hardlinks are usable using `beets.util.hardlink`.
-
-    The `beets.util.hardlink` function catches EXDEV errors and raises
-    `FilesystemError`.
-    """
-    with tempfile.TemporaryDirectory() as tmpdir_str:
-        tmpdir = Path(tmpdir_str)
-        src = tmpdir / "hardlink_src"
-        dst = tmpdir / "hardlink_dst"
-
-        try:
-            src.touch()
-
-            util.hardlink(util.bytestring_path(src), util.bytestring_path(dst))
-
-            return dst.exists() and src.stat().st_ino == dst.stat().st_ino
-
-        except (FilesystemError, OSError) as e:
-            log.debug(f"Hardlink check failed: {e}")
-            return False
-
-
-@functools.lru_cache(maxsize=1)
-def check_reflink() -> bool:
-    """Tests if reflinks are usable using `beets.util.reflink`.
-
-    The `beets.util.reflink` function catches EXDEV and EOPNOTSUPP errors and
-    raises `FilesystemError`.
-    """
-    with tempfile.TemporaryDirectory() as tmpdir_str:
-        tmpdir = Path(tmpdir_str)
-        src = tmpdir / "reflink_src"
-        dst = tmpdir / "reflink_dst"
-
-        try:
-            # Bytes are needed to test copy-on-write (COW) behavior
-            src.write_bytes(b"test")
-
-            # Attempt the reflink. `fallback=False` ensures it fails if not supported,
-            # instead of silently performing a full copy.
-            util.reflink(
-                util.bytestring_path(src), util.bytestring_path(dst), fallback=False
-            )
-
-            return dst.exists() and src.stat().st_ino != dst.stat().st_ino
-
-        except (FilesystemError, OSError) as e:
-            log.debug(f"Reflink check failed: {e}")
-            return False
-
-
-# Run/Test feature probes
-HAVE_SYMLINK = check_symlink()
-HAVE_HARDLINK = check_hardlink()
-HAVE_REFLINK = check_reflink()
-
-
-class InputError(Exception):
-    def __init__(self, output: str | None = None):
-        self.output = output
-
-    def __str__(self) -> str:
-        msg = "Attempt to read with no input provided."
-        if self.output is not None:
-            msg += f" Output: {self.output!r}"
-        return msg
-
-
-class DummyOut:
-    encoding = "utf-8"
-
-    def __init__(self) -> None:
-        self.buf: list[str] = []
-
-    def write(self, s: str | bytes) -> None:
-        if isinstance(s, bytes):
-            s = s.decode(self.encoding, "replace")
-        self.buf.append(s)
-
-    def get(self) -> str:
-        return "".join(self.buf)
-
-    def flush(self) -> None:
-        pass
-
-    def clear(self) -> None:
-        self.buf = []
-
-
-class DummyIn:
-    encoding = "utf-8"
-
-    def __init__(self, out: DummyOut | None = None):
-        self.buf: list[str] = []
-        self.reads: int = 0
-        self.out = out
-
-    def add(self, s: str) -> None:
-        self.buf.append(f"{s}\n")
-
-    def close(self) -> None:
-        pass
-
-    def readline(self) -> str:
-        if not self.buf:
-            if self.out:
-                raise InputError(self.out.get())
-            else:
-                raise InputError()
-        self.reads += 1
-        return self.buf.pop(0)
-
-
-class DummyIO:
-    """Mocks input and output streams for testing UI code."""
-
-    def __init__(self) -> None:
-        self.stdout: DummyOut = DummyOut()
-        self.stdin: DummyIn = DummyIn(self.stdout)
-        self._orig_stdin: TextIO | None = None
-        self._orig_stdout: TextIO | None = None
-
-    def addinput(self, s: str) -> None:
-        self.stdin.add(s)
-
-    def getoutput(self) -> str:
-        res = self.stdout.get()
-        self.stdout.clear()
-        return res
-
-    def readcount(self) -> int:
-        return self.stdin.reads
-
-    def install(self) -> None:
-        self._orig_stdin = sys.stdin
-        self._orig_stdout = sys.stdout
-        sys.stdin = self.stdin
-        sys.stdout = self.stdout
-
-    def restore(self) -> None:
-        if self._orig_stdin is not None:
-            sys.stdin = self._orig_stdin
-        if self._orig_stdout is not None:
-            sys.stdout = self._orig_stdout
-
-
-class AssertionsMixin:
-    """A mixin with additional unit test assertions."""
-
-    assertions = unittest.TestCase()
-
-    def assert_exists(self, path: Path) -> None:
-        """Assertion that a file exists."""
-        assert path.exists(), f"file does not exist: {path!s}"
-
-    def assert_does_not_exist(self, path: Path) -> None:
-        """Assertion that a file does not exists."""
-        assert not path.exists(), f"file exists: {path!s}"
-
-    def assert_equal_path(self, path_a: Path, path_b: Path) -> None:
-        """Check that two paths are equal. This resolves relative paths and symlinks, so
-        it checks if the paths point to the same location on disk.
-        """
-        path_a_full = path_a.resolve()
-        path_b_full = path_b.resolve()
-
-        assert path_a_full == path_b_full, (
-            f"paths are not equal: {path_a!s} and {path_b!s}"
-        )
-
-
-# A test harness for all beets tests.
-# Provides temporary, isolated configuration.
 class TestCase(unittest.TestCase):
     """A unittest.TestCase subclass that saves and restores beets'
-    global configuration. This allows tests to make temporary
-    modifications that will then be automatically removed when the test
-    completes. Also provides some additional assertion methods, a
-    temporary directory, and a DummyIO.
+    global configuration.
+
+    Deprecated: migrate tests to use the ``beets_plugin_env`` fixture instead.
     """
 
     def setUp(self) -> None:
-        # A "clean" source list including only the defaults.
         config.sources = []
         config.read(user=False, defaults=True)
 
-        # Direct paths to a temporary directory. Tests can also use this
-        # temporary directory.
         self.temp_dir = Path(tempfile.mkdtemp())
 
         config["statefile"] = str(self.temp_dir / "state.pickle")
         config["library"] = str(self.temp_dir / "library.db")
         config["directory"] = str(self.temp_dir / "libdir")
 
-        # Set $HOME, which is used by confit's `config_dir()` to create
-        # directories.
         self._old_home = os.environ.get("HOME")
         os.environ["HOME"] = str(self.temp_dir)
 
-        # Initialize, but don't install, a DummyIO.
         self.in_out = DummyIO()
 
     def tearDown(self) -> None:
@@ -264,3 +64,16 @@ class TestCase(unittest.TestCase):
         self.in_out.restore()
 
         config.clear()
+
+
+__all__ = [
+    "HAVE_HARDLINK",
+    "HAVE_REFLINK",
+    "HAVE_SYMLINK",
+    "PLATFORM",
+    "AssertionsMixin",
+    # "DummyIn",
+    # "DummyOut",
+    # "InputError",
+    # "TestCase",
+]
